+++ front/src/app/register/register.module.css
@@ -5,7 +5,7 @@
   height: 100vh;
 
   /* Fondo monocromático con degradado */
-  background: linear-gradient(135deg, #1a1a1a, #444444);
+  background: radial-gradient(circle at 0 0, #2a3a7a 0%, #181a2a 40%, #181a2a 100%);
 }
 
 .card {
